update dir context
